craftersbackend/requirements.txt
djangorestframework-simplejwt
dj-rest-auth[with_social]
django-allauth